backend/requirements.txt
Flask==1.1.4
Flask-Cors==4.0.0
Flask-Sockets==0.2.1
gevent==23.9.1
gevent-websocket==0.10.1
MarkupSafe==1.1.1
psycopg2==2.9.9
python-dotenv==0.19.1
PyJWT==2.8.0
